Fix matrix row and column mixup
m_matrix elements point to the rows, so the first index should be
related to the height, and the second index to the width.

It does end up working without this fix, because the matrix width is
constrained to be 1 in both vectorMagnitudeSquared() and dot(), and the
matrix elements are stored consecutively, in one array, so the memory
accesses are exactly the same in both cases. So, this is mostly a
cosmetic fix (ATM), but can help with understanding the code.